+++ Models/scripts/train-model4.R
@@ -1,9 +1,16 @@
 #
 #   Model 4: Random forest
 #
+#   AB: mark runtime
+#
 
 # to ID output files
 model_prefix <- "mdl4"
+
+library(here)
+library(lgr)
+
+setwd(here::here("Models"))
 
 # Load needed packages and setup global variables controlling model training
 source("scripts/0-setup-training-environment.R")
@@ -70,18 +77,19 @@ if (file.exists("output/models/mdl4_tuning_results.rds")) {
 
 # Parameter search space
 ps_rf = makeParamSet(
-  makeIntegerParam("num.trees", lower = 200L, upper = 700L),
-  makeIntegerParam("mtry", lower = 10L, upper = 40L)
+  makeIntegerParam("mtry", lower = 10L, upper = 50L),
+  makeNumericParam("sample.fraction", lower = 0.7, upper = 1)
 )
 
 # Random forest with hyperparameter optimization via 10-fold CV over random grid
 auto_rf <- makeTuneWrapper(
-  learner    = makeLearner("classif.ranger", predict.type = "prob"),
+  learner    = makeLearner("classif.ranger", predict.type = "prob",
+                           num.trees = 1000, min.node.size = 10),
   par.set    = ps_rf,
   resampling = makeResampleDesc("CV", iters = TUNE_CV_FOLDS),
   measures   = list(mlr::brier),
   control    = makeTuneControlRandom(maxit = RANDOM_TUNE_SAMPLES),
-  show.info  = FALSE)
+  show.info  = TRUE)
 
 
 learner <- auto_rf
@@ -89,38 +97,24 @@ task    <- full_task
 
 
 #
-#   Cross-validation
-#   ________________
-
-# Get OOS performance of model
-mdl_resamples <- resample(learner, task, resample_strategy,
-                          measures = ACC_MEASURES)
-
-# Estimate final version of model on full data
-mdl_full_data <- mlr::train(learner, task)
-
-# Training predictions for full data
-preds_oos <- tidy_ResampleResult(mdl_resamples)
-
-# Save artifacts
-write_rds(mdl_resamples, path = sprintf("output/models/%s_resamples.rds", model_prefix))
-write_rds(mdl_full_data, path = sprintf("output/models/%s_full_data.rds", model_prefix))
-write_rds(preds_oos,     path = sprintf("output/predictions/%s_cv_preds.rds", model_prefix))
-
-
-#
 #   Test forecasts
 #   _______________
 
 
+lgr$info("Running test forecasts")
 test_forecasts <- test_forecast(learner, task, TEST_FORECAST_YEARS)
-
 write_rds(test_forecasts, path = sprintf("output/predictions/%s_test_forecasts.rds", model_prefix))
 
 
 #
 #   Live forecast
 #   ________________
+
+
+lgr$info("Running live forecast")
+# Estimate final version of model on full data
+mdl_full_data <- mlr::train(learner, task)
+write_rds(mdl_full_data, path = sprintf("output/models/%s_full_data.rds", model_prefix))
 
 forecast_data <- split_data$fcast %>% select(-!!TARGET)
 fcast <- predict(mdl_full_data, newdata = as.data.frame(forecast_data))
@@ -146,4 +140,5 @@ parallelStop()
 #   model_prefix variable set at the beginning of this script.
 #
 
+lgr$info("Models done, sourcing assessment script")
 source("scripts/assess-model.R")