+++ Models/scripts/train-model5.R
@@ -5,6 +5,11 @@
 # to ID output files
 model_prefix <- "mdl5"
 
+library(here)
+library(lgr)
+
+setwd(here::here("Models"))
+
 # Load needed packages and setup global variables controlling model training
 source("scripts/0-setup-training-environment.R")
 # import functions
@@ -41,38 +46,24 @@ task    <- full_task
 
 
 #
-#   Cross-validation
-#   ________________
-
-# Get OOS performance of model
-mdl_resamples <- resample(learner, task, resample_strategy,
-                          measures = ACC_MEASURES)
-
-# Estimate final version of model on full data
-mdl_full_data <- mlr::train(learner, task)
-
-# Training predictions for full data
-preds_oos <- tidy_ResampleResult(mdl_resamples)
-
-# Save artifacts
-write_rds(mdl_resamples, file = sprintf("output/models/%s_resamples.rds", model_prefix))
-write_rds(mdl_full_data, file = sprintf("output/models/%s_full_data.rds", model_prefix))
-write_rds(preds_oos,     file = sprintf("output/predictions/%s_cv_preds.rds", model_prefix))
-
-
-#
 #   Test forecasts
 #   _______________
 
 
+lgr$info("Running test forecasts")
 test_forecasts <- test_forecast(learner, task, TEST_FORECAST_YEARS)
-
 write_rds(test_forecasts, file = sprintf("output/predictions/%s_test_forecasts.rds", model_prefix))
 
 
 #
 #   Live forecast
 #   ________________
+
+
+lgr$info("Running live forecast")
+# Estimate final version of model on full data
+mdl_full_data <- mlr::train(learner, task)
+write_rds(mdl_full_data, file = sprintf("output/models/%s_full_data.rds", model_prefix))
 
 forecast_data <- split_data$fcast %>% select(-!!TARGET)
 fcast <- predict(mdl_full_data, newdata = as.data.frame(forecast_data))
@@ -98,4 +89,5 @@ parallelStop()
 #   model_prefix variable set at the beginning of this script.
 #
 
+lgr$info("Models done, sourcing assessment script")
 source("scripts/assess-model.R")